backend/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
ecdsa==0.19.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
passlib==1.7.4
# Google Cloud 관련 추가
google-cloud-sql-connector[pg8000]
pg8000
google-auth
google-cloud-core
google-cloud-storage>=2.10.0
